Moved replaceTitleWithSpanned and replaceSpannedWithTitle to combined module called convert-title. #28
